ai-agent/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-text-splitters
pypdf
pydantic
pydantic-settings
python-multipart
faiss-cpu
python-dotenv
openai
